+++ src/event_poller.cc
@@ -40,7 +40,8 @@ void EventPoller::
     }
 }
 
-EventPoller::EventPoller() : _epoll_fd(epoll_create1(_EPOLL_CREATE1_FLAGS)) {
+EventPoller::EventPoller()
+    : _epoll_fd(epoll_create1(_EPOLL_CREATE1_FLAGS)) {
     if (_epoll_fd == -1) {
         LOG_SYS_FATAL << "epoll_create1 failed";
     }
@@ -113,8 +114,12 @@ void EventPoller::poll_for_active_events_of_all_fds(
 
     // epoll_wait might be interrupted by signal handlers, so make a loop for it
     while (true) {
-        current_event_array_size =
-            ::epoll_wait(_epoll_fd, _event_array, _MAX_SIZE_OF_EVENT_ARRAY, -1);
+        current_event_array_size = ::epoll_wait(
+            _epoll_fd,
+            _event_array,
+            static_cast<int>(_MAX_SIZE_OF_EVENT_ARRAY),
+            -1
+        );
 
         if (current_event_array_size == -1) {
             if (errno == EINTR) {
